chore: extract constants
