+++ src/feature/area-chart/chart.ts
@@ -12,9 +12,9 @@ import { addHours } from "date-fns";
 import EventEmitter from "eventemitter3";
 
 import { ScaleLinear, ScaleTime } from "../../types";
+import { AXIS_HEIGHT, AXIS_WIDTH } from "../../util/constants";
 import { Zoom, ZoomTransform } from "../../util/zoom";
 import { Data } from "./area-chart";
-import { AXIS_HEIGHT, AXIS_WIDTH } from "./constants";
 import { Contents } from "./contents";
 import { Colors } from "./helpers";
 import { UI } from "./ui";
