+++ src/feature/area-chart/contents.ts
@@ -5,8 +5,8 @@ import { range } from "../../helpers";
 import { Renderer } from "../../renderer";
 import { Container } from "../../renderer/display";
 import { ScaleLinear, ScaleTime } from "../../types";
+import { AXIS_HEIGHT, AXIS_WIDTH } from "../../util/constants";
 import { SeriesData } from "./chart";
-import { AXIS_HEIGHT, AXIS_WIDTH } from "./constants";
 import { HorizontalGrid, LineCurve, VerticalGrid } from "./display-objects";
 import { AreaCurve } from "./display-objects/area-curve";
 import { Colors } from "./helpers";
