+++ src/feature/line-chart/chart.ts
@@ -5,8 +5,8 @@ import { addHours } from "date-fns";
 import EventEmitter from "eventemitter3";
 
 import { ScaleLinear, ScaleTime } from "../../types";
+import { AXIS_HEIGHT, AXIS_WIDTH } from "../../util/constants";
 import { Zoom, ZoomTransform } from "../../util/zoom";
-import { AXIS_HEIGHT, AXIS_WIDTH } from "./constants";
 import { Contents } from "./contents";
 import { Colors } from "./helpers";
 import { Data } from "./line-chart";
