--- src/feature/line-chart/constants.ts
@@ -1,14 +1,0 @@
-/**
- * Standard font size in CSS pixels
- */
-export const FONT_SIZE = 12;
-
-/**
- * Height of the bottom date axis
- */
-export const AXIS_HEIGHT = FONT_SIZE + 5;
-
-/**
- * Width of the vertical price axis
- */
-export const AXIS_WIDTH = FONT_SIZE + 60;
