+++ src/feature/line-chart/display-objects/vertical-axis.ts
@@ -5,7 +5,11 @@ import { InteractionEvent } from "../../../renderer/interaction/interaction-even
 import { Text } from "../../../renderer/text";
 import { ScaleLinear } from "../../../types";
 import { Colors } from "../../../components/depth-chart/helpers";
-import { AXIS_HEIGHT, AXIS_WIDTH, FONT_SIZE } from "../constants";
+import {
+  AXIS_HEIGHT,
+  AXIS_WIDTH,
+  FONT_SIZE,
+} from "../../../util/contants/constants";
 import { Gesture } from "../zoom/gesture";
 import { Zoom } from "../zoom/zoom";
